+++ source/RTSCamera.CommandSystem/src/Orders/VisualOrders/RTSCommandActivateFacingVisualOrder.cs
@@ -46,7 +46,7 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
             orderToAdd.OrderType = _orderType;
             if (orderToAdd.OrderType == OrderType.LookAtDirection)
             {
-                if (IsFromClicking && Patch_OrderTroopPlacer.IsFreeCamera && CommandSystemConfig.Get().OrderUIClickable && CommandSystemConfig.Get().OrderUIClickableExtension)
+                if (IsFromClicking && Patch_OrderTroopPlacer.IsFreeCamera && CommandSystemConfig.Get().OrderUIClickable)
                 {
                     // Allows to click ground to select target to facing to.
                     OrderToSelectTarget = SelectTargetMode.LookAtDirection;